Introduce AddressPrefix struct.
This struct is used for defining the kind and the intended network of
an address.  It also provides static conversions methods to and from
various different address kinds and intended blockchains.

This logic is being moved outside of the BlockChainIdentity class to
support more complex address prefixes as well as proving a cleaner
boundary between the Bitcoin protocol code and wallet abstractions.