Filter relevant transaction outputs with correct indexing.
Previously btcwallet performed this filtering, but I changed the API
to remove the output index from each of the reported outputs and
instead decided that all outputs should be included.  This caused a
bug in Paymetheus where the IEnumerable index for the Select LINQ
expression referenced the incorrect transaction output.  This was
fixed by changing btcwallet to return details regarding every output.

Now that every output is included in the RPC responses, filtering must
be done on the client side to remove outputs that we don't care about.
Indexing is fixed since the lengths of the transaction outputs are
equal for both the decoded transaction itself and the additional
wallet details provided by the TransactionDetails message.