Add method to import seed when creating a wallet.
This only adds the dialog and will call the CreateNewWallet RPC with
the decoded seed rather than a randomly-generated one.

As btcwallet does not support recovering addresses and transactions
for an imported seed, this feature won't work correctly and
transaction history will have to be recovered after generating the
same number of addresses used from the previous wallet.  However, once
that is implemented, this GUI feature should work as intended.

The seed can be inputed using either hex or the PGP word list
encoding.  At the moment, no checksumming is used or removed from the
input, and all bytes are considered part of the seed.  If any
additional data is added to a seed encoding later, the
WalletSeed.DecodeAndValidateUserInput method will need to be changed
to interpret this separately.

Closes #24.