Refactor address pool code and change database storage to account index
The old address pool code was faulty in that it could not scale to
more accounts than the default account and did address index alignment
using an actual address instead of an index. This updates the address
pools to be generalized so that in the future they can be multiaccount.
Address indexes for an account are now stored in the meta bucket of the
address manager database. The logic for the address pool has been
greatly simplified.