Use Paymetheus' local AppData directory for wallet.
Add AssemblyResources static class for reading attributes from the
entry assembly.  This is used to access the name and organization of
the application which is used for the local appdata directory.

Modify the Portability.LocalAppData static method to take an
organization name.  When empty, only the product name is used,
otherwise folder for the product is nested under an organization
directory.

Fixes #52.