Look up application title from resources.
Set the full window title earlier in the application lifetime, instead
of appending an alternate network name for non-mainnet wallets after
the wallet has been opened and synchronization completed.

Rename the title from just "Paymetheus" to "Paymetheus Bitcoin
Wallet".  This will allow forks for other currencies to rename the
title so multiple Paymetheus instances can be run without confusion
about which currency each is for.  This affects both the window title
and the process name in the Windows task manager.