Add optional resyncing options to importscript and importprivkey
New optional resyncing options have been added to importscript and
importprivkey for convenience. A rescan height can be included as the
last argument so that the user can easily rescan from a height not
the genesis block. The code for importscript was also refactored so
that it may easily be called from the new RPC server. glide locks
were updated to newer versions of dcrd and dcrwallet.