Hide the Border for the generated address unless one was generated.
Simplify the UI by always showing the "generate address" button.
Click again to generate a new address, replacing the old.

Improve the xaml so it doesn't look like it got generated by 15
different automated tools.

Make the address manually selectable so it can be copied, since the
label under the address is lying and it is not automatically copied to
clipboard yet.

Fixes #13.