Implement transaction history.
The transaction history view is now filled with valid information
about transactions relevant to the selected account, ordered from
oldest at top down to newest at the bottom. Transactions include the
total debit/credit amount to the account, ignoring change (spending a
10 DCR input and receiving 5 DCR back as change will show the debit as
-5 DCR rather than -10).

Note that rendering large numbers of transactions will be very
slow. This is an unfortunate side effect of a complex DataTemplate
being used in a ItemsControl/ListBox/ListView, rather than using a
DataGrid. This gives us the ability to expand transaction details, but
a more efficient view will need to be implemented in a later release.

Fixes #9.