Update overview xaml to present recent transactions.
The recent txs were already being added and updated in the viewmodels.

There's a lot more that can be done here, such as expanding selected
transaction items to show more details.

Originally there were plans that selecting an item would replace the
overview with an entirely new view devoted to showing all of the
transaction's details.  Right now I think this disrupts the user
(especially since this is the only place in the app that we'd need
some kind of "back button").  Removing this view for now
(OverviewDeeper.xaml) and this can be figured out later.  What's more
important now is to put some moderately useful data here instead of
having a giant overview page with nothing but 2 numbers at the top.

Fixes #77.