Fix calculation of transaction debit/credit amount to the wallet.
Only consider whether a transaction output is controlled or not when
determining whether to add it as a credit.  This fixes the calculation
when transactions are sent to an external address from the same wallet
(e.g. when moving funds across accounts).

Debits should be negative and credits positive, so fix the signs.

Fixes #136.