Protect access to the shared Wallet structure.
Because both the UI thread and the wallet RPC synchronization tasks
may access the wallet structure at the same time, access to the
structure must be synchronized.  Add a new Mutex<T> type which when
unlocked, provides access to the underlying locked instance (in this
case, the Wallet).  This ensures that (unless the wallet reference is
leaked outside of the exclusive access) all access to the wallet only
occurs when the lock is held.  This is also preferable to acquiring a
private mutex in each of the Wallet's public APIs as it allows
multiple public APIs to be called without any other tasks from
changing wallet state inbetween calls.

The mutex can only be acquired asynchronously.  This ensures that the
UI thread never blocks attempting to access wallet state.

Fixes #114.