Update project dependencies.
Add 'version: master' to all btcsuite and decred projects, since we
always want the latest versions of these.  This works around a bug in
the latest version of glide where it does not track the master branch
correctly.

Update gRPC for the 1.0.0 release and update bolt to 1.3.0 to pick up
a fix for a memory handling error that was exposed by the Go 1.7 SSA
compiler backend.

Fixes #325.

Fixes #328.