Process transactions atomically with connected blocks.
Save all block headers and ensure that they attach.  With all headers
saved, the wallet is able to perform more validation checking when
attaching blocks to the main chain.  This also improves compatibility
with SPV, which requires all block headers to be saved.

Significantly improve the rescan logic so it is more understandable
and reusable, and in the future, cancelable.  This also leads the way
to add a new RPC to trigger a rescan starting from some block hash or
height.

This change includes a database upgrade, so it is not possible to use
an updated database changed by this commit on older versions of the
wallet.

Fixes #315.