Add controlled startup RPCs to the gRPC interface.
This allows the client to orchestrate the bringup and syncing of the
wallet with the consensus RPC server.  This does add complexity to the
client, but the tradeoff is the client is well aware of what the
wallet is doing and can present more detailed information to the user.
This can improve the user experience.

This change also adds the RPC for BIP0044 address (and optional
account) discovery.  Account discovery is one of the new features this
API provides, and is required for clients to perform seed restoring.

Closes #385.

Closes #389.