Remove cmd/dropwtxmgr and doco references to it.
This tool is no longer necessary to force rescans with the new rescan
RPCs available in both the JSON-RPC and gRPC servers, and the new
versions are better anyways (can start the rescan at any block height,
and for the gRPC version, the RPC returns a stream of progress
notifications).

For situations where problematic transation data must be removed from
the wallet database, a reseed should be performed.  It is better to
have a single method of fixing these wallets instead of several.

The wtxmgr namespace is now required to exist when opening a wallet.
Older versions of the dropwtxmgr tool should not be used on the
database anymore.

Closes #433.