ticketbuyer: add dynamic max price default, remove highpricepenalty queue (#501)
The idea behind highpricepenalty was presumably to reduce volatility, 
but in practice it did not work as we can see see from the ongoing high 
volatility. It was also misaligned with user incentives since it would only
 cause people to still buy overpriced tickets. It seems most people 
were just disabling it. Also bad was that the code was very buggy and 
cumbersome. So, it is being removed. A simple dynamic max price 
setting is being added to be sure to protect users from buying 
expensive tickets when they are just running the default config. With 
this, we can move forward and start looking at other possible ways to 
reduce volatility that work, while keeping user profits optimized.

This also fixes the problem caused by the ticket queue where all funds
 available were not being used. It will now buy as many tickets as 
possible which is the expected behavior.

Closes https://github.com/decred/dcrticketbuyer/issues/57