Always commit database transaction when buying tickets. (#618)
This resolves double spend errors that are caused by creating and
publishing multiple transactions while in a single database
transaction.  If any of the tickets can not be purchased, the
transactions that were published remain tracked by the database.

The PurchaseTickets method has also been changed to always return the
hashes of all successfully published tickets, even if less than the
requested number could be bought.  This allows ticketbuyer to update
its statistics with the number of bought tickets if only some were
successfully sent.

While here, add a comment about why a hardcoded minconf=0 in the auto
ticket buyer code is needed.