Prevent panic in JSON-RPC server.
This fixes panics caused by dereferencing a nil wallet to determine if
an RPC that requires the --unsafemainnet flag when mainnet is the
active network.  The fix is to pass in the active network parameters
to the RPC server and use those rather than trying to access them
through the nil wallet.