Add the actual votebits to the gRPC APIs.
The stakepool API will require setting the actual votebits.  Include
these in the results to make this easier for wallet clients.

Note that it's not strictly necessary to include these since clients
could bitwise OR the bits themselves, but it is more efficient to
return the bits directly to them, especially after setting choices.