Remove a silly error from accountsyncaddressindex.
The accountsyncaddressindex RPC only extends watched addresses now and
doesn't change any actual index that the account is synced to.
Therefore there is no reason to bother returning an error when a child
index behind the last used address is passed.

This was caught because an account branch without any address use
(lastused=^uint32(0)) was hitting an error when any valid child index
was being passed.

While here, add an additional error to prevent trying to sync to child
indexes for hardened keys.