Upgrade created watching only DBs to latest version.
This fixes a panic that was caused by creating watching only databases
marked with the latest DB version when they really were in fact the
initial version of the udb database schema.

Watching only wallet databases that were created using the faulty code
(such as that found in the 1.0.0 and 1.0.1 releases) will not open
correctly due to the incorrectly-recorded database version and must be
recreated.