Use correct confirmation check for ticket expiry.
Tickets expire after they have 40960 + 256 + 1 confirmations (on
mainnet).  Unfortunately, the confirmation check used by the wallet
was not taking into consideration the extra confirmations required
before the ticket becomes live.  Fix this so it matches consensus
rules.

The consensus rules calculate and compare the height that the ticket
expires at rather than checking the total number of confirmations (the
depth of the transaction in the blockchain), but wallet uses
confirmation checks everywhere so this was kept consistent with the
rest of the wallet code.