Fix PoolAddress and TicketAddress options with new AddressFlag.
Use embedded Address instead of the *AddressFlag when assigning to Address fields of ticketbuyer.Config.  This compiled because *AddressFlag satisfies the dcrutil.Address interface.  However, this broke later tests against nil because it was itself not nil (whether nil interface{}, nil *AddressPubKeyHash, etc.), but a non-nil pointer to the an AddressFlag.