Update Decred and golang.org/x deps.
This change also includes a fix for an txscript API break which
removed the StandardVerifyFlags constant.  Wallet now defines this
constant itself, based on the newly introduced
mempool.BaseStandardVerifyFlags.  As features are enabled, these flags
should be updated.  An even better solution, but not one that can be
performed at the moment, would be for the wallet to know when a
voted-on fork activates and change the flags automatically.