Don't set heights for manually added tickets.
The height field of a pool ticket saved by the stake manager was never
being used.  To remove a case of requiring the dcrd RPC client to
query for its height, this field is now left zeroed and a comment has
been added to the field indicating that it is not used and can not be
relied on.

If the height of the transaction is required to be known, it must be
looked up through the transaction manager.