Concurrently discover account branch used addresses.
The previous code would search accounts concurrently but both branches
would be scanned sequentially.  In best case scenarios, this can halve
the amount of time performing the address discovery.

This change also uses the errgroup package to wait for all of the
tasks to run and merge the errors, as this is simpler and less error
prone than the error channel previously used.