Switch default coin types to those defined in SLIP0044.
This change modifies the wallet to prefer the use of the coin types
assigned to Decred in SLIP0044 over the coin types used thus far.  New
wallets that are created with a random seed are created using the
SLIP0044 coin type, and if no address usage from the legacy coin type
is found during address discovery, the wallet is automatically
upgraded to use the SLIP0044 coin type instead.

This change introduces a database upgrade and older versions of the
wallet will not work on upgraded wallet DBs.