grpc: add pagination support to GetTransactions
This changes the behaviour of the WalletService.GetTransactions grpc
call, from creating an in-memory array to streaming the transactions as
they are fetched from the db.

It also adds a new parameter to the call, target_transaction_count to
limit the amount of transactions returned and allow clients to better
paginate transaction listing.