Remove the sendtosstx JSON-RPC method.
This method is not recommended for use since the addition of the
purchasetickets RPC.  While there may be a use case for creating a raw
ticket purchase, this is already possible using the raw transaction
APIs and does not need a dedicated RPC just for ticket purchases.

I searched public code across all GitHub repos and did not find any
usage of this RPC.

The JSON-RPC API version has been bumped to 5.0.0 to reflect the
breaking change.