Fix getbalance totals.
Totals should be included when '*' is used to query all accounts, not
when querying a particular account.  The totals in the previous result
were not the totals of all accounts, but just copies of the balance
values of the the single queried account.

While here, remove float64 summing and sum the dcrutil.Amounts
instead to prevent loss of precision.