Fix check for unset public passphrase.
The default config value for the public passphrase is "public" not the
empty string.  Fixing this check will correctly log a message to set
the public passphrase with the appropiate config options.

While here, edit the log message to remove -- from the flag names, as
these are more appropiate to add to the config file anyways.