Allow queries of txs spending wallet outputs.
The wallet database records which transactions and inputs spend wallet
outputs, but this information was not querably through the wallet's
internal APIs or the RPC servers.

A new gRPC method WalletService.Spender is added which queries the
transaction and input index of any transaction (unmined or in the main
chain) which spends an output relevant to the wallet.

Spenders of outputs which are not relevant to the wallet can not be
queried through this API.