Add experimental vgo support.
With the vgo proposal accepted for integration with future versions of
the Go toolchain, we have decided to add experimental module support
to dcrwallet.  vgo provides several desirable properties over
vendoring, such as removing the need for projects to be compiled under
$GOPATH/src and the ability to `vgo get` a module to fetch and build a
project with all required dependencies in a single command.

At this time, dep remains the recommended way to build and manage
dependencies for dcrwallet.  The module files do not interfere with
dep but allow us to experiment with vgo.  This provides a way to
determine whether or not vgo meets our needs so we can determine if
and when to switch.