Prevent signrawtransaction panic on gettxout result.
The dcrd gettxout JSON-RPC method is invoked by signrawtransaction to
look up previous output scripts that the wallet may not record itself.
When gettxout is queried for an output that is already spent in the
main chain, it will return the JSON null without any error.  This
requires special casing in the signrawtransaction implementation to
avoid a possible nil pointer dereference.