Make walletdb and drivers internal packages.
To limit the public API surface before defining modules, the walletdb
package is being made an internal package so it may not be imported by
other packages outside of the wallet directory.

The walletdb.DB type was previously passed around by other packages in
the code, so to make this change possible, an opaque DB wrapper type
(wallet.DB) has been introduced.  This type only exports a single
method of the io.Closer interface, but allows the wallet package to
access the internal interface definition.