various: Updates for proper syncing via grpc for version 5.0.0
This allows dcrwallet to be synced via requests over grpc in normal mode.

Added FetchMissingCFilters request, because committed filters are required 
for both spv and normal mode. 

RescanPoint is a new request that provides the last place the wallet had been 
rescanned. This allows for more efficient rescanning and discovering addresses 
requests.

DiscoverAddressRequest now allows for a starting_block_hash to be set, which
works well in conjunction with the newly added RescanPoint request.

The existing GetTickets wallet function has been renamed to GetTicketPrecise.
The current GetTickets function does not have a chain client argument which 
means it is no longer able to determine whether unspent tickets have been missed.