Avoid ticketbuyer deadlock
When the v2 ticketbuyer processes a block notification, it should not
simultaneously block reading of the next block notification.  This causes
deadlocks as code which writes to the subscribed tip-changed client is blocked
by an ongoing ticket purchase.

To prevent this, ticket purchases are run in their own goroutine and do not
block the read of the next block connected notification.