Revert RPC address discovery to prior algorithm
When address/account discovery is performed in the RPC sync mode, it must
support features used by voting wallets, such as discovering usage of addresses
as voting addresses in ticket purchases.  These output scripts are not included
in the committed filters by design (to make SPV voting impossible), but without
them it is not possible to correctly restore a voting wallet.

Due to this limitation of using committed filters for address discovery, the
prior algorithm which depends on the dcrd address exists index is being
reintroduced and is used when the Peer implementation is the RPC network
backend.