Buy max tickets mineable per block.
The previous logic limited the total number of buyable tickets per block to the
number of picked tickets per block.  This is not the desired behavior, and the
buyer should buy as many tickets as can be mined in the next block but no more.

Noticed by @alexlyp who had to switch back to the v1 buyer temporarily on
testnet3 as the v2 buyer was not buying enough tickets to support the target
ticket pool.