Remove dep files
A switch to Go modules from dep has been made for the 1.4 development
cycle.  Go modules now provide reproducible builds for the dcrwallet
executable (same functionality as dep) but additionally provide
multiple units of versioning ("modules") in a single repo, as opposed
to the repo itself being the unit of versioning (dep restriction).
The executable build is also able to take advantage of submodules
defined in dependency projects, such as dcrd.  As these submodules are
not on identical version schedules as the dcrd executable, dcrwallet
will no long need to track the dcrd master branch, and there will be
greater build compatibility for projects importing both dcrd and
dcrwallet for library packages.

With the switch to modules, the dep files (Gopkg.toml and Gopkg.lock)
will become stale very soon, and in some ways already have (a dep
vendored build still works, but consider how the imported siphash
package used still has correctness issues on ARM processors).

Because it's an unnecessary burden to expect all developers to manage
and keep dependency files updated correctly for both versioning
schemes, and because the existence of the files indicate dep support
to others looking at the code, it's better to simply remove these and
rely solely on modules.