Fix maturity off-by-one calculations
The wallet was previously treating maturity as the required number of
block confirmations when checking for coinbase, ticket, and
revocation/vote output maturity, as well as for ticket expiry.
However, maturity describes the number of blocks after a tx is mined
in which must exist before the output is spendable.  Because the block
the tx is mined in is not considered in this calculation, but is
considered one block confirmation, an off-by-one bug existed where
outputs were calculated as matured one block before they were.

This fixes this off-by-one by adding helper functions to return
whether a particular type of transaction output is matured given the
tx and tip height, as well as tests exercising these functions.