rpc: Add RpcSync rpc command
This command allows for syncing the wallet to a connected consensus
server via rpc.  It requires that the user supplies proper rpc connection
credentials and a valid certificate.  Once connected the wallet will
provide notifications in the same for as SpvSync.  Fetching headers and
missing cfilters progress, discovering addresses and rescanning notifications
will be sent to the stream.  After all the steps are complete, a final
Synced notifications will be sent.