[release-v1.3] Return errors for details of unknown transactions.
When a transaction is not yet known, lookups for details of the tx
must return an appropriate error.  This commit modifies the internal
TxDetails function to return non-nil errors rather than nil details
structs if the transaction is not recorded.

This is not a breaking change to consumers of the wallet module as the
udb package is a candidate for becoming internal and is not possible
to use in other projects as an internal walletdb interface is
necessary to create a udb.Store.

This is a backport of 58cf828c.