[release-v1.3] multi: Fix GetTickets on non-voted (#1289)
This fixes two small issues related to the GetTickets grpc call:

- The error was not correctly returned in GetTickets/GetTicketsPrecise,
causing it to not bubble up the stack
- The udb function Store.TxDetails was recently changed to return an
error when a transaction is not found; this kind of error needs to be
ignored in TicketDetails when loading the voting details of a ticket
that hasn't yet voted.

This is a backport of 39d3bc0.