wallet: Track individual ticket commitments
This commit changes the handling of ticket commitments so that the
wallet can correctly account for its share in tickets instead of always
considering it has access to the full funds.

It aims to fix balance issues by decoding the commitment addresses in
tickets processed by the wallet and individually tracking those only if
the address belong to the wallet.

This introduces database changes and therefore upgrades the wallet udb
to version 12.