[release-v1.4] Fix reorgs of blocks with 2+ tx spend chains
If a reorganized block contains two wallet transactions, one which spends a
relevant wallet output from the previous transaction, the previously removed
mined credit value must have been recorded in memory for correct spend tracking
while moving the values to unmined buckets.  This map was not being assigned the
removed value, causing the lookup of the previous credit to fail with a "missing
credit value" error.