notifications: Fix race condition
This fixes a race condition detected by external tests using the `-race`
switch. The condition is fixed by using the notification server's mutex
to fully protect the currentTxNtfn member.

When sending an aggregated tx notification, only the current outstanding
events (at the time of the start of the sendAttachedBlockNotification
call) will be sent to subscribed clients.