wallet: Ignore duplicate tx error when publishing
This makes the wallet not purge unmined transactions when the underlying
error returned by the backend is a dcrjson error with ErrRPCDuplicateTx
code.

Previously, this caused wallets to purge an unmined transaction that was
republished or otherwise already existed in the network even though it
was still valid, and could cause the wallet to produce a double spend
transaction by marking the funds of the (now purged) tx as spendable.