Update to latest x/crypto for Salsa20 fix
This includes a fix for Salsa20 on amd64 platforms when over 256GiB of the
keystream is consumed.

Salsa20 is used by nacl, which is in turn used by dcrwallet for encrypting
values in the database, but it is improbable that 256GiB of the keystream would
ever be consumed by any wallet usage.

See https://golang.org/issue/30965 for more details.