Fix cursor indexes to derive next address.
The cursor index was not being set to the correct value, causing addresses to be
returned multiple times after any block or unmined transaction was received.
This has been fixed and tests have been added to ensure the behavior remains
correct, as this is not the first time address reuse has become a problem.