Replace wallet locking management goroutine with mutexes.
This change simplifies the management of the wallet's locked and unlocked states
when multiple concurrent requests may lock, unlock, or change the passphrase.
It also increases concurrent use of the unlocked wallet as requests to call
unlock with a valid passphrase do not block on current operations which make use
of the unlocked wallet.

The Start/Stop/WaitForShutdown methods are no longer necessary following this
change and have been deprecated.