Add context to JSON-RPC method handlers.
This change adds support for the context.Context type throughout the JSON-RPC
method implementations, allowing for request-scoped cancelation.  The context is
canceled whenever a Websocket or HTTP POST client is seen disconnecting.

Requests are not canceled whenever the process receives an interrupt signal, but
this may be implemented at a later time.