Add wallet address interfaces.
The following new interfaces are introduced:

- V0Scripter: for addresses used to create version 0 scripts
- SecpPubKeyHasher: for addresses created from a secp256k1 pubkey hash
- SecpPubKeyer: for addresses with a known secp256k1 pubkey
- SecpPubKeyHash160er: for addresses paying to a secp256k1 pubkey HASH160
- BIP44AccountXpubPather: for BIP0044 addresses derived from an account xpub
- AddressP2PKHv0: a union interface for v0 P2PKH addrs
- BIP44XpubP2PKHv0: a union interface for v0 P2PKH addrs with known xpubs

These interfaces describe the behavior and features of an address and allow
callers to operate on their distinctive capabilities without switching on a
concrete type.  More interfaces are planned to be introduced in the future as
address and script usage evolves.

The wallet methods NewExternalAddress, NewInternalAddress, and NewChangeAddress
have been modified to return addresses implementing all of the new interfaces.