Move JSON-RPC server implementation to jsonrpc package
When the gRPC server was first added to btcwallet, it was done with
the intention of eventually replacing all usage of the JSON-RPC
server, due to API issues inherited from bitcoind, the lack of any
named parameter support, and the lack of request cancelation in
JSON-RPC.  However, there are no such plans to deprecate or remove the
JSON-RPC server anymore.  This package rename reflects that decision.

This is not considered a major module bump.  Non-main packages in the
main module are not for external consumption and we have stopped
tagging the main module with versions recognized by the module
tooling, at least until the main module only contains a main package
and possibly internal packages.