Generate new block locators after announced blocks.
When a block is announced and it is attached to and becomes the new
best chain, the previous block locators saved by the SPV syncer are
cleared.  This could cause an slice out-of-bounds panic in getHeaders
since that method required that the locators be at least length 1 at
all times.  This change protects against that logic race by always
checking whether a new locator generation has zero length and if so,
new locators are generated and saved.