Replace rpcclient/v2 and introduce wallet/v3
This commit replaces rpcclient/v2 with github.com/jrick/wsrpc/v2,
which is much simpler in both implementation and API surface, as well
as being easier to use with arbitrary RPCs with parameters and result
types marshaled with encoding/json.  Due to the use of rpcclient/v2 in
the wallet's public package API, the wallet module and all modules
also using it in their public API have been bumped to their next major
version.

A new module, github.com/decred/dcrwallet/rpc/client/dcrd is
introduced.  This module provides a client for type-safe convenience
methods to perform JSON-RPC with dcrd, and is created from a minimum
(single method) interface.  Wallet methods take this same interface as
a parameter and create the client as needed, which reduces coupling
between the JSON-RPC API and the wallet's public package API.

Because it has been so long since the last breaking change to the
wallet module, it is likely that it will not be tagged at v3.0.0 for
some time.  This will let us make breaking changes in tree before we
are satisfied with the latest API.