Use decred.org/dcrwallet as main module name
This modifies packages of the main module to begin with the
decred.org/dcrwallet prefix.  This is a long overdue change since the
conversion to modules technically should have caused a /v2 module to
be created, but due to other project decisions, it was not possible to
bump dcrwallet's major version to 2.  Renaming the module has the same
desired effect, however.

Because this change creates a new (and unrelated) module to the
previous github.com/decred/dcrwallet module, this opportunity is used
to move all packages from the main module which are not versioned or
intended for external consumption under the internal directory.

Because this module shares the same git repo and history as
github.com/decred/dcrwallet, versioning will begin at 1.5.0.

To simplify this conversion process, the
github.com/decred/dcrwallet/internal/zero module has been discarded.
It is easier to copy the zero functions where needed rather than
maintaining an internal package across many modules with different
prefixes.

This change will eventually allow the latest release of dcrwallet to
be installed with go get with the simple command:

  $ go get decred.org/dcrwallet