Do not cleanly unload wallet during panic
A panic was observed deep in the bbolt module due to database file
corruption after a system crash.  Running the wallet with this
corrupted database would cause an internal consistency error in bbolt
and a panic inside of a database update.  During unwind, the bbolt
mutex was not released.  Attempting to cleanly unload the wallet in
this state caused a deadlock.  Instead, rethrow the panic when
panicing so that the full stack trace can be observed and the wallet
process to exit.