Add assets package for accessing dcrwallet file assets
The assets package is part of the main module.  It provides a single
function, Path, which returns the absolute path to any file in the
dcrwallet main module, relative to assets package.  This function is
usable by release tooling to access module file contents without
needing to copy the contents into a Go variable or constant.