Switch to Version 2 CFilters
This switches the wallet to use version 2 committed filters.

Version 2 cfilters are consensus enforced to be comitted to in block
headers after activation of DCP0005, therefore they are safer for SPV
clients (including the wallet) to use.

This introduces database migration v15 which drops the existing (v1)
cfilters from the wallet db, forcing a re-download of the missing
cfilters. The wallet is modified to verify post-DCP0005 cfilters against
the corresponding block header, which ensures the filter is valid
according to consensus rules.

Pre-DCP0005 filters are validated by a new json-rpc call
(validateprecdp005cf) and eventual problems may be fixed by running a
repair tool (cmd/repaircfilters) which imports the hardcoded and
known-good filters into the wallet.