Update version prerelease and build metadata naming scheme
Releases will now use the empty string for the prerelease field, and
"release.local" for build metadata.  This indicates that the code is
for a release, but the release was built from the repo or using go get
without the release build tool.

Development versions will use "pre" for the prerelease and default to
no build metadata.  The following commit will update the fields for
continued development of master.