Fix panic on next attached block following importxpub
While the database was being properly updated for import xpub
accounts, and the decryption errors for the missing nonce have been
corrected when no encrypted private key was recorded, the wallet would
eventually die in a panic on the next attached block following an
importxpub RPC invocation due to a nil pointer dereference of the
(missing) in-memory account xpub.  This change corrects this bug by
adding an account buffer for the imported account (instead of only
being read on next dcrwallet startup from the database) and watches
the initial gaplimit number of children in both BIP0044 branches.