Avoid spurious error logs caused by cancelling ticket purchasing
When the context passed to the ticket purchasing methods is cancelled,
an error log was being created for the failed deferred call to
watchHDAddrs.  Rather than continuing to call this method, if the
context is cancelled, return early from the deferred function.

Because the automated ticketbuyer uses context cancellation to stop
ongoing purchases whenever the ticket price changes, this was causing
unnecessary and unhelpful errors to be logged.