Define Address and KnownAddress interfaces
Address is a generic address type, which may or may not represent an
address or key managed by the wallet.  Unlike dcrutil.Address, this
interface defines the method to produce the script bytes and version,
whereas dcrutil.Address cannot create a script and instead the
txscript package must be used (but this operation does not work for
every possible dcrutil.Address implementation).

KnownAddress is an Address with additional methods providing details
regarding the wallet account associated with the address.  A new
wallet method, also named KnownAddress, replaces the previous AddrInfo
method which returned an udb type.

Three other address interfaces are defined which add to the methods of
the KnownAddress interface.  These are:

* PubKeyHashAddress: a KnownAddress for any P2PKH output script.

* P2SHAddress: a KnownAddress for any P2SH output script. This type
  requires the redeem script to be known.

* BIP0044Address: a KnownAddress for any BIP0044 account address
  (including imported xpub accounts).

All existing wallet methods continue to operate on dcrutil.Address,
but it is expected that these will be cut over to use wallet's defined
Address interface in the future.  Compatibility wrappers will likely
be required for some time to implement the interface for a parsed
dcrutil.Address.

If a dcrd package ever copies the definition of the Address interface,
the wallet's definition can become a type alias to this interface
without breaking any backwards compatibility.

The ideas for address interfaces were originally implemented in commit
958ad8bfb9d3171bdeab838d19534a97e6676341.  The new interfaces simplify
and replace these previous interfaces.  The reduced complexity is
largely due to addresses always knowing how to encode as a transaction
script, rather than needing to define this method on top of an
embedded dcrutil.Address.

In addition to the Address types, the concept of an account kind or
purpose is also introduced.  These are defined by new AccountKind
constants in the wallet package.  Currently, there are three account
kinds (for seed-derived BIP0044 accounts, for the pre-created imported
account, and for imported xpub accounts).  It is planned to introduce
more account kinds as well, e.g. accounts which derive P2SH multisig
addresses using incrementing BIP0044 keys.