Switch DetachedBlocks ntfn to report header
This switches the notification server to report the full block header of
detached blocks instead of just the hash.

This is more flexible for callers which need more information from a
detached block and would not be able to obtain it from the wallet given
that after the notification is triggered the detached blocks are rolled
back from the wallet db.

This also modifies the appropriate gRPC endpoint to return this
information. Additionally, it adds the previous block hash to the
BlockDetails gRPC message.