Remove more udb types from public wallet API
This is done by defining types idenical to the old udb.Balances and
udb.AccountProperties in the wallet package, and modifying udb's
definitions to be type aliases of an unnamed struct type.

While here, clean up some method names and change the return on
AccountBalances to a slice instead of a map.  The map served no
purpose and the only caller simply created a sorted slice from the map
values anyways.