Add capability to save per-ticket vote preferences
Add optional `tickethash` parameter to the `getvotechoices`, `setvotechoice` json-rpc
methods and `VoteChoices`, `SetVoteChoices` gRPC `VotingService` methods.

If either `setvotechoice` methods is called with the `tickethash` parameter set, the
specified voting preferences will be recorded for that ticket alone, otherwise the
specified voting prefs would be saved to the default preference store. Per-ticket
voting preferences override voting preferences in the default store when voting.

Similarly, if either `getvotechoices` method is called with a `tickethash` parameter,
the individual choices for that ticket are returned. If no ticket-specific choices
were previously saved for the ticket, the default preferences are returned.