Avoid uncommon CoinJoin output amounts
When large UTXOs are mixed, the mixed output amount was chosen as the
smallest preprogramed amount which would result in one or more mixed
outputs.  This had the unfortunate side effect of advertising seldom
used mixed output amounts to a CoinShuffle++ server which were
uncommon relative to amounts advertised by clients participating in
mixed ticket buying.

As a workaround, the current sdiff is queried, and a decision is made
whether to use an even smaller mixed amount which will allow for
better pairings.