Upgrade database account format
A database upgrade is introduced which rewrites all BIP0044 accounts
into a new serialization format. The account "row" is modified to only
contain the information that never changes (such as the account's
extended pubkey and privkeys) and a "variables" buckets are created
for each individual account to record, in separate key/value pairs,
account fields that do change over time (such as the account name and
last used/returned child indexes). This allows faster access to read
and modify the particular field(s) in question, rather than requiring
all account metadata to be read or written back to the db when only a
small amount of data must be changed.

Upcoming PRs will take advantage of the account variables buckets to
record additional properties of accounts in a way that retains
compatibility with existing account management.