also use blake2b keyed hashes for the manager's global passphrase
the encrypted accounts use this mechanism now, so keep it the same.
this is much better than simple concatenation of a salt onto the
passphrase and using sha512, which is vuln to length extension
attacks.  If this were to keep sha2 at all, this should have be using
a hmac construction.