Move address definitions to payments package
The wallet.NetworkBackend interface is now able to use dcrwallet's own
Address interface rather than using dcrutil in the definition.  This
required moving the interfaces and several implementations to be moved
from the wallet package to a new payments package to avoid a package
import cycle.

The old wallet.KnownAddress interface is renamed to wallet.Address
since it is no longer confused with the previous wallet.Address
interface (which becomes a payments interface).