p2p: Bump max supported pver
This bumps the maximum supported protocol version to the upcoming
version 8, which adds the getinitialstate/initialstate pair of messages.

It also fixes a bug that could happen if the wire package was upgraded
to a version with a new protocol version but the wallet's p2p package
was not updated: in that case SPV clients could negotiate an incorrect
version and fail in undefined ways.

This is fixed by ensuring the protocol version advertised by the wallet
is always the maximum one the wallet understands as opposed to the
default one for NewMsgVersion().