Make passphrases optional in most gRPC methods
Since the client is now authenticated using TLS client certificates,
it is safe to allow gRPC methods to be called and use wallet private
keys even if the request itself does not provide the private
passphrase.  For backwards compatibility, keep the existing behavior
when a nonzero passphrase is provided, but do nothing to unlock the
wallet or check the passphrase when it has zero length.

In order for this to be useful, a future commit must add the ability
to lock or unlock the wallet from separate methods.  This
functionality already exists for individually-encrypted accounts.