Update to latest gRPC release
The current protobuf packages were bumped previously due to other
updates to golang.org/x modules, and additional gRPC methods cannot be
added until the new autogenerated services are compatible with our
gRPC version.  Bump the version to the latest release so that upcoming
gRPC methods to lock and unlock the wallet may be added.