Add gRPC WalletService methods to lock/unlock wallet
These methods will allow the wallet to be persistently unlocked for a
period of time, during which other gRPC methods which use wallet
private keys may be performed without necessarily providing the
passphrase in each request.

While here, add the missing documentation for the
UnlockAccount/LockAccount methods, and bump the API minor version.