Allow gRPC and JSON-RPC auth types to be separately specified
We need to not necessarly require TLS client cert authentication for
JSON-RPC simply because the gRPC server is running and requiring it.
This change removes the --authtype flag, replacing it with only a
--jsonrpcauthtype flag to change the behavior of the JSON-RPC server.
TLS client cert authentication remains required for gRPC, and requires
the presence of a file containing trusted client certificates.

While here, improve logging to warn when RPC listeners are configured,
but client certs are enabled and no trusted client certs exist to
authenticate any connections.