Handle VSP's successful processing of fee payment
When the VSP has already received a fee payment as valid, the wallet
should update its status for the ticket when trying to process the
payment again on a later restart or due to a RPC request from
Decrediton.  This was not being done, and instead the API error
returned by vspd for this situation was being silently ignored by the
SyncVSPFailedTickets gRPC method and never logged, because that method
was erroneously assuming the error would always be logged elsewhere
(nevermind that dcrwallet logs are not the method by which RPC clients
should obtain error information).