Randomize inputs selected for paying VSP fees
The input selection was not randomized previously (as this code did
not use the randomized InputSource that was previously added for input
selection for all funded transactions).

Randomizing these chosen inputs may (randomly) allow VSP-using ticket
purchasing to succeed where it would have reliably failed before, if
the user's high-valued UTXOs are not selected for fee payments.  This
would then enable a user or a client to reattempt the purchase.