Fix JSON-RPC error message when sending from locked wallet
This error no longer returned a helpful error message about what to
run ever since the per-account encryption was implemented, because
walletpassphrase was not in all cases the correct method to run if the
account was encrypted by unique keys.  Fix this error message to
suggest either walletpassphrase or unlockaccount, and wrap
errors.Locked errors when the account is locked and created from an
xpriv, not an xpub.