[release-v1.6] Promote error kind even if wrapped error would be overwritten
This fixes an issue where a wrapped errors.Error can lose its Kind
matching when another error object is used in the E function.  There
is still the issue of the earlier error being forgotten dropped
entirely, but this at least fixes most detection issues that depend on
the Kind matching.

A test (which fails without the patch) is added to demonstrate the
issue.