[release-v1.6] vsp: Rewrite client for persistent bg processing
This change involves almost a complete rewrite of the vsp client
package, removing the syncing goroutines and instead placing methods
of a feePayment structure on a timer at randomized intervals to
perform each step with the VSP.

One design goal of this change is to ensure that all interactive vsp
client usage is performed on a feePayment structure that is also being
background processed.  This allows fee payment processing to continue
even if the VSP client was temporarily created and not actively
syncing new tickets.

Another separate goal of this rewrite is to allow the client to query
the VSP for tickets that have been processed by the VSP, but are not
recorded as such in the wallet's database.  This will be used to
recover fee payment information for tickets after seed restores.