Handle VSP ticket status "received".
The "received" status is expected during normal operation. The wallet should not just log a warning, it should handled this status the same was as the "broadcast" status - continue waiting for the status to update to "confirmed".

Without this handling, the wallet never notices that the fee has been confirmed and never removes the ticket from the VSP client.