Remove unused vsp types
One of these types was still using json.RawMessage for the echoed
request, which will no longer work with vspd, as it will send the
message base64 encoded as a string. There is no need to backport this
to the release-1.6 branch to fix this type, because the struct was not
used at all.