[release-v1.6] Include remaining UTXOs for additional split during low balance
If there was not enough remaining balance to reserve outputs to pay
for one ticket (after reserving outputs for paying a VSP fee), but the
sum of the already-reserved UTXOs and the remaining UTXOs in the
account do sum to enough value to create the additional split
transaction, that split transaction should still be created.  Not
including these outputs in the additional split was causing confusing
"insufficient balance" errors for users, as not all of the UTXOs in
the account were being considered.

For example, consider a ticket price of 100 and an account with two
UTXOs of 60.  One output would be reserved for paying the fee, but
reserving additional outputs to pay for a single ticket would fail,
and this second output of 60 was not being included in the logic for
creating the additional split transaction.  By including this output,
even when it is not enough value to pay for the ticket alone, the
additional UTXO split will be created.