Move package-global VSP client map to loader package
Add to the map when the VSP client is created for the ticket autobuyer
with the global application settings.

This will be used by an upcoming change to the JSON-RPC server to
allow voting preferences to be set and updated through this server.
Currently, it is only possible to modify VSP voting preferences for
registered tickets through the gRPC server.