Derive correct public/private key for encrypted accounts
When an account is individually encrypted, whether or not to derive
the private key instead of the public key depends on whether the
account is unlocked, rather than querying the unlocked state of all
other accounts without individual encryption keys.

While here, remove a pointless break statement and an unused struct
field.