Don't attempt to vote/revoke without the private key
This error is commonly seen on ticketbuyer wallets which will attempt
to revoke any missed tickets, even though an importing voting xpub
account prevents access to the private keys to do so.  This is not an
error, and should be detected up front before ever constructing the
unsigned vote/revocation.

When the private key is recorded by the wallet, but the wallet or
account is locked, the wallet will still error when failing to sign
the revocation.