vsp: Add VSP Host and PubKey info to udb
This diff upgrades the udb to now include vsp host and pubkey
information for all vsp ticket entries.  This is done by
utilizing 2 new buckets for host and pubkey.  The host bucket
is keyed by a simple counter and the pubkey is keyed by
host url.  The vsp tickets now have a VSP host id field.
If they don't yet have a VSP attached they are set to 0.

This diff also adds a 'confirmed' state for vsp tickets.  This
is used once a vsp has confirmed with the client that the fee
has been confirmed and voting will proceed on the ticket.
Once confirmed, the vsp ticket will no longer be processed
during ProcessManagedTickets.