Protect passphrase MAC hasher with mutex
Using the hasher mutates its state, but it was being used by the
UnlockedWithPassphrase method with the manager's read mutex held.
This adds an additional mutex for protecting just the hasher.

This fixes a data race caused when concurrently unlocking the wallet
from several goroutines, and the invalid passphrase errors that were
caused by racing on the hasher.