Use last returned key counts in grpc Accounts response
The last returned counts are more useful as they indicate the true
number of keys derived for an account, rather than the last one that
has been marked as used.

While here, remove the hardcoded addition of 20 for the gap limit.
The gap limit is configurable so this hack was incorrect.

A later commit can add additional fields to this message to include
the last address index marked used, under a more appropriate field
name.