Remove sync mechanism which held wallet unlocked
This mechanism has not worked properly ever since accounts have been
able to be individually encrypted with a separate passphrase, as it
only applied to the accounts covered by the wallet's global
passphrase.  It also appears to be the cause of at best a bad UX and
at worst a deadlock whenever an unlocked mixing wallet is locked.