spv: Improve reorg handling
This fixes an issue where the SPV syncer could fail to handle reorgs if
it only had a single peer to connect to. This can usually happen during
automated tests of SPV wallets, when the wallet is directed to a simnet
peer.

This is fixed by making sure the peer re-requests the missing headers in
case it receives announcements for orphan headers.

This also refactors and removes some code that is now no longer used.