peering: Fix ping pong hang.
Don't block in writeMessages when an error result cannot be sent. Ignore
the errors and let garbage collection take care of the go function.

Group context detecting blocks in one function of errgroup. Disconnect
with that error after return.

Also error with timeout if the ping portion of a ping pong cannot be
sent after a time limit.