wallet: Fix cfilter missing on 1-deep reorg
When a sidechain header was previously received via a header
announcement msg but then the cfilter was fetched via a getheaders msg
with a different peer, the filter was not filled to the blocknode due to
a missing check on the duplicateNode() function to account for the root
of the sidechain forest.

This could cause issues with 1-deep reorgs due to the missing cfilter
later on when switching to the new sidechain as a best chain.

This fixes the issue by also testing the root of the sideforest for
duplicates and filling the fetched cfilter as needed.