db: Increase version to 25.
importVotingAccount is the 25th version of the database. This version
indicates that importing a new account type has been enabled. This
account facilitates voting using a special account where the private key
from indexes of the internal branch are shared with a vsp. The new type
is not recognized by previous wallet versions. This version only updates
the db version number so that previous versions will error on startup.