vsp: don't set/use a default policy
This avoids the possibility of using a previously created vsp client
that has a policy different from the one intended by the caller.

In most cases already, the caller has a policy defined that should be
used rather than some previously set default policy. The only other
case where a policy to be used isn't readily handy is in the json-rpc
handler for the processUnmanagedTicket method. That method uses the
vsp client set in config and should thus use the policy set in config.