rpc: Correct JSON-RPC and gRPC verifymessage.
This corrects the main logic that verifies messages to create a p2pkh
address from the recovered public key instead of a p2pk since p2pkh is
what is produced when signing and therefore what is required when
verifying (as can be verified in dcrd as well).

This also updates the call sites that decode the addresses to properly
reject p2pk addresses for the same reason.