Prevent panic in RPC-mode getstakeinfo
The expired slice was accessed out-of-bounds due to dcrd no longer
supporting the existsexpiredtickets method.  Avoid these RPCs entirely
(including existslivetickets) by instead counting missed or expired
tickets by the block heights of their revocations.  (This is not
necessarily 100% accurate for tickets that were missed prior to the
activation of DCP0009, but there is no better way for wallet to
provide this result without tracking the entire ticket pool.)