Implement importpubkey JSON-RPC method
This method imports compressed 33-byte secp256k1 public keys and their
derived P2PKH addresses to the wallet's imported account.

This method may only be used by watching-only wallets.  The rationale
for this limitation is that there is only one account that can deal
with the importing of arbitrary keys, and UTXO selection is unaware of
whether private keys are available to spend a UTXO or not.  This
requires that accounts must not contain some addresses with private
keys available and others where only the public key is available.

This is a backport of aa43a44a8e60d986f1fac7c30498793eb90d5010.