Add checkpointing and enforce testnet3 hard fork
In order to follow the intended chain after the unvoted testnet3 hard
fork, we implement checkpointing and require a block from the fork to
be present in the main chain.  This is performed by rollbacking the
main chain at startup if an invalid checkpoint is detected, and
refusing to extend the main chain unless a checkpoint rule is
satisfied.  We use checkpoints to push wallet onto the intended
unvoted chain due to it having significantly less total work than the
chain being reorged.

This is a backport of c8bae9634ced9b82d13bd2709407476afb940628.