Avoid context cancellation race
In the event that a purchase tickets request context from the ticket
autobuyer is cancelled due to the changing sdiff window, addresses and
outpoints must still be watched.  Not doing so eventually leads to an
invalid UTXO set as relevant transactions become missed.

There is a small chance that this change may cause a hang during
process shutdown.  A better solution for this is included in a
comment.

Backport of c6e7b5f9af7711f5ea2476faac95094d4003c043.