Implement faster random UTXO lookup
This uses cursors seeking to a random key to randomly pick from UTXOs
in the bbolt db.  Previously, the randomization was implemented by
reading all UTXOs into memory and shuffling.

To avoid endless spinning, random attempts are only performed for half
of the total number of elements in the bucket, after which the
algorithm reverts back to reading everything again to shuffle the
remaining results.  For this to be possible, a new method is added to
the database implementation returning the number of key/value pairs in
a bucket.

Backport of 42b602c9444879ad2f913329c4e45a2c363d82fd, excluding API
changes like the removal of the unused UnspentOutputsForAmount
function and the addition of KeyN method to the DB interface.