Always avoid creating small mix change
The mix with the smallest denomination never creates change; instead,
the extra value is rolled into the transaction fee.  Do the same for
all other mixing (ticket splitting, and the larger denominations of
account mixing) so that the unmixed account is not filled with these
outputs that are too small to mix, and too tempting to spend together.

Backport of dc82294b976a09ba6d1b7f22099e54454dc4e8b0.