Respect proxy config for SPV and seeder connections
Remove the address of the local peer from the handshake.  This is to
prevent leaking internal network details and if a proxy is in use.
Peers are also no longer informed of their external address if the
address is loopback or a private or shared address.

Backport of 18d4c023835206e455b65a71df82df8fc28255b6.