multi: Bump major module version to 4
This increases the major version of the decred.org/dcrwallet module to
version 4. This is needed to prepare for future changes which will be
breaking changes from a semver point of view.

All references to packages in the decred.org/dcrwallet/v3 module are
changed to their respective v4 versions.