udb: Remove v25 upgrade test
This test only verifies the database version number and not any real
functionality. It is inconsistent with the other upgrade tests which do not
perform this check.

The test was also brittle because it assumed the post-upgrade database to be
exactly version 25. This has not failed so far because the assumption happens to
be correct, but it will fail in future when newer DB versions are created.