vsp: Dont remove payments unless confirmed by VSP.
The only way to be certain a VSP has acknowledged a fee tx and added a
ticket to its voting wallets is to check FeeTxStatus == "confirmed".
Checking for 6 confirmations on the fee tx is not an adequate
substitute.

For example, if a client has broadcast it's own fee tx rather than
waiting for the VSP to do it, it's possible that the tx could be mined
and have 6+ confs, but the amount paid is not enough, or the payment is
sent to the wrong address.