Log transaction changes for non-initial rescans
For all rescans except for the initial one when a wallet is first synced
(during reseed, or after process restart), log those transactions that are
changed status from unmined to mined in a block, and any newly-observed
transaction that is added to the wallet.

These later rescans are typically done to fix a problem syncing blocks and/or
transactions, and these transaction hashes should be useful information to
discover what previously went wrong.