[release-v1.8] vsp: Hardcode background ctx
The VSP client code was not handling contexts correctly. It was
receiving short-lived contexts bound to the lifespan of a single RPC
request, and using them to create feePayment instances. Most of the code
in feePayment executes async, so should be running on a long-lived
context such as context.Background().

Backport of 19f7ed1ef8ca7e9b38d178b7dc2cee73df18c071