vsp: Recalculate ticket live/expiry heights.
Previously the live & expiry heights were only being set once upon
feePayment initialization, which meant that an unmined ticket added to
the VSP client would have heights set to zero forever.

With this change, if either of the values are found to be zero, we check
if the ticket has been mined yet and the heights are recalculated to
their correct values if so.