[release-v1.8] vsp: Recalculate live/expiry height
Previously the live & expiry heights were only being set once upon
feePayment initialization, which meant that an unmined ticket added to
the VSP client would have heights set to zero forever.

With this change, if either of the values are found to be zero, we check
if the ticket has been mined yet and the heights are recalculated to
their correct values if so.

Backport of 87228676c4e98ec520b790a10fcc1a51199f6b33