spv: Move FetchMissingCFilter to syncer startup
This moves the "fetching missing cfilters" stage of syncing from
individual peer startup to the syncer startup.

This makes it possible to use multiple peers to fetch cfilters from
(which is not yet implemented) and reduces overall load by ensuring this
stage is only done once instead of being done for each new peer.