spv: Move addr discovery and rescan to syncer startup
This moves the address discovery and rescan procedures from each
individual peer startup to the global syncer. This allows using multiple
peers for the rescan and removes the need for having a syncer global
lock, since it is now guaranteed that there is a single execution of the
rescan procedure.