spv: Check all peers for initial sync height
This modifies the SPV syncing logic to check all peers for any that have
an advertised height larger than the current wallet tip height to use as
a syncing peer.

This ensures that all peers are checked for headers during initial sync,
instead of a single one.

It also adds a new function to disconnect from peers that have fallen
behind the wallet's mainchain tip height during the initial sync,
ensuring the wallet does not hold onto a peer that can't possibly help
the syncing process.