spv: Avoid re-hashing header
This makes a change to the block node initialization code to, instead of
calculating the block hash of each header, use the PrevBlock member of
the header when multiple headers are available.

This is safe to do because the lower P2P level has already validated
that a received header batch connects in sequence (and therefore the
PrevBlock of a header actually is equal to the prior header's block hash).

This avoids unnecessarily performing duplicate work in calculating the
block hash and reduces overall cpu usage, specially during the initial
sync process.