spv: Track requested block locator height
This makes the wallet BlockLocators function return the height of the
first (i.e. most recent) block locator.

This makes checking if the peer failed to provide headers independent of
the state of the db.

While this makes no functional difference at the moment, in the future,
when fetching headers and updating the db from different goroutines, this
makes it easier to reason that the check is being done correctly.