spv: Refactor initialSyncHeaders
This moves all the header fetching logic from the old getHeaders
(renamed to initialSyncHeaders) to a separate getHeaders function.

In the future, this will make it easier to refactor the initial sync
header fetching to perform this stage in parallel to the other stages.

This is mostly a code moving commit.