p2p: Reduce number of goroutines in CFiltersV2()
This refactors the CFiltersV2 function to reduce the number of
goroutines used, therefore slightly reducing total cpu and memory load
and making it easier to run the race detector and debug issues.

The number of goroutines is reduced from one per block hash in the
request to two per call.  This is done by refactoring the code to use
one goroutine to send all requests and one to read the responses, while
ensuring that the requests are still sent as fast as the network stack
will allow it.