chain: Avoid early getHeaders on sidechains
This adds a check for the presence of the parent block on a known
sidechain prior to re-requesting all headers on announced blocks during
RPC chain sync

Prior to this commit, the parent block was checked for presence only on
the main chain, which could cause getHeaders to be repeatedly executed
on reorgs of size greater than one.

This was mostly an issue for automated tests that generate large reorgs.