spv: Request headers from peer after initial sync
This modifies the peer setup procedure to request any updated headers
from a newly connected peer if the connection is made after the initial
sync is completed.

This speeds up re-syncing after initial sync when the local client has
been offline for some time (for example, due to a flicking network
connection) by ensuring any and all updated headers are requested
immediately, instead of waiting until a new block is announced to begin
the catchup process.