wallet: Fix off-by-one in addr discovery
This fixes an off-by-one in sizing the backing array for cfilter data
during address discovery.

During address discovery, the GetMainChainCFilters call to fetch
cfilters from the DB uses an array sized by the caller to put the data.
However, due to being an inclusive fetch, the call in the address finder
is not correctly sized, missing one (i.e. the last) block.

This could cause an issue when the wallet had a single address used and
that address was used on a transaction on the wallet's tip, causing the
wallet to miss that address being used.