wallet: Remove mainchain tip check on NeedsAccountsSync
This check could prevent a wallet from going through account discovery
if it had not completed the initial chain sync prior to being restarted.

One instance this could happen is when restoring SPV wallets under an
instable network connection or having a power loss during the initial
sync of a restored wallet.