wallet: Allow coin type upgrades on simnet
This removes the enforced coin type upgrade for simnet wallets, in order
to allow testing the upgrade code path under simnet.

Prior to this commit, the coin type of a restored simnet wallet was
always upgraded to the corresponding SLIP0044 coin type, in order to
ensure that the address meant to be used as the mining address for the
dcrd node was never made invalid. However, despite this enforced
upgrade, there is still a bug with account and address discovery if the
wallet was started before at least blocks 1 and 2 were mined in the
underlying full node.

To alert simnet users about this bug this commit adds a note displayed
during simnet wallet creation.

In the future, this bug may be fully fixed by ensuring the underlying
node is fully synced and has blocks before the sync process completes.
This will need to take into account use cases where the underlying node
is meant to be fully unsynced (such as air-gapped wallets).