multi: Introduce --offline sync mode
This adds the --offline sync mode to the wallet. This mode allows the
wallet to run without connecting to the network through either SPV or
RPC modes. Calls that do not require access to the network can still be
executed.

This is mostly useful for running wallets which are meant only to
generate addresses and sign transactions without actually tracking
balances, such as air-gapped wallets.

Previously, such wallets required running with an unconnected dcrd
instance or in SPV mode with --spvconnect specified to an address that
would not provide network services, but these methods are prone to be
misused and recent changes have broken their correct behavior.

The offline mode is introduced by writing a custom NetworkBackend
implementation that is guaranteed to never attempt any external
connections.