Improve error message for initial csppsolver check
There was some confusion over the `csppsolver` option that resulted in it
being set as if it were a boolean option.  This caused a properly built and
working csppsolver to not be used, and instead, the solverrpc package
attempted look up and exec a file named "1".  Adding the error message would
have removed some confusion as this was debugged.