Record derived mixing change addresses
These database updates were being appended to a slice of updates to run in the
future, but no such updates ever took place.

Since this is only done here for change addresses, and only one address is
derived a time from this method, remove the deferred update batching and
instead just record the derived address in the DB immediately.