[release-v2.0] chain: Wait for errgroup before returning from sync
If syncing over JSON-RPC is being reattempted in a loop, as the main package
does, not waiting for the errgroup that runs the wallet's mixing client may
cause it to concurrently enter mixclient.Client.Run again before it has
returned.  This eventually results in a panic due to closing the already-closed
warming chan.

Backport of 8b9a63d81e3849aed46daad367cbd52d9d4f9de1.