[release-v2.0] chain: Fetch unknown PRs of notified orphan KEs
This changes the notification handler when a run-0 KE is received but the
sender's own PR is not yet known to request the PR from dcrd.

Because this also works to initially sync the mixpool during the warming
period, the previous call to getmixpairrequests is no longer necessary and has
been removed.

Backport of 9dbfe06584d68c19c11022b80575195b8d545d23.