p2p: Ban peer with mix inv when disablerelaytx is active
Version 2.0.1 of dcrd will now not send invs related to mixing when
disablerelaytx is signalled during the peer negotiation stage.
Therefore, the wallet should ban peers that incorrectly send such invs
when the wallet was configured with this flag.

While there may still be some peers running v2.0.0 (which would cause
them to be banned from the wallet), their number should be small enough
to not cause significant connection issues.