Attempt cxns to >=3 mixing-capable peers
Similar to dcrd, we want SPV wallets to maintain a minimum number of peers
that support the mixing protocol version.  This will be especially important
after dcrwallet is updated to a mixclient branch that submits PR messages over
the entire duration before the epoch occurs, as any errors for failed message
sends due to not having a single mixing-capable peer connected would be
delayed far into the future.