main: Update ticket buyer startup logging.
In all code and log messages the ticket buyer is referred to by name,
however in startup/shutdown logging it was given the more vague name
"auto transaction creator".

Its also helpful to log whether or not the ticket buyer is purchasing
mixed tickets because that has a large impact on its behaviour and
potential failure modes.