Dont allow ticket voting address override.
Setting a specific voting address was an important part of purchasing a
legacy stakepool ticket, however it is no longer necessary and providing
these options only serves as a potential footgun which can lead to
accidental address reuse.

Now a new voting address is derived for every ticket, either from the
voting account if one is specified, or from the purchase account if not.

Breaking changes:

1. Config `--votingaddress` removed.
1. Param `ticket_address` removed from `PurchaseTickets` gRPC.
1. Param `voting_address` removed from `RunTicketBuyer` gRPC.
1. Param `ticketaddress` removed from `purchaseticket` json-RPC.