[release-v2.0] ticketbuyer: Dont terminate if RescanPoint fails.
If the call to RescanPoint fails the ticketbuyer should not be
terminated forever, it can continue running and subsequent calls to
RescanPoint may succeed. This is how all other errors in the ticketbuyer
are handled (with the exception of incorrect passphrase error).

Backport of 197810606e8dee6fa65a1a161f2f9d814f31919b.