[release-v2.0] Always use configured voting account when mixing
The comments suggested that this was the intended behavior, but the voting
account option was only being considered when UseVotingAccount (allowing the
option to be used in non-mixing situations) was true, and mixed ticketbuying
was not setting this value.

Backport of 17dd44862ff2a8e0b4086bd528aca172f43de5bd.