[release-v2.0] Add context wrapping for syncer disconnections
Any long-lived process (such as mixing) must return if the network backend
(e.g. the dcrd RPC connection) is disconnected or the syncer encounters
another error.  However, it was observed that mixclient calls were continuing
to execute despite the syncer erroring, resulting in locked outputs that could
not be mixed after the syncer reconnected and restarted.

To avoid this, the network backend interface gains additional methods to
select on when disconnect occurs, and a new context wrapping function is
added, which creates a derived context that is canceled (or "done") after the
syncer exits.  Wrapped contexts are added in the ticket purchasing and account
mixing paths, including those being performed by the autobuyer.

Backport of 4adf892651e1ffcdbeaa8f5cacc26e59dae04285.