Check that consolidate output is not dust
One of consolidate's uses is to sweep up existing dust.  It would be quite
ironic for it to create an additional dust output.

Adding this dust check also causes consolidate to error sooner when the
estimated fee is greater than the total input sum, preventing creating a
transaction output with a negative amount.