Prevent VSP fee rate overpayment
When an extra output is required to be created to pay the VSP during UTXO
contention, the estimated fee tx size included an unnecessary change output
and a buffer of 2x the estimated network fee.  However, this would create
change outputs small enough to be considered dust, causing the additional
value to be rolled into the network transaction fee instead, causing a roughly
3x network fee overpayment for these VSP fee transactions.

Fix this by removing the unnecessary change output from the size calculation
and do not add more than 1x the estimated network fee for the 1:1 VSP fee
transaction.