Ignore repeat blockconnected notifications
dcrd was observed to send blockconnected notifications for the same block
multiple times, apparently during a series of reorgs.  If we already have the
notified block, return out from the blockconnected notification handler
without also recording the block as a sidechain.  This prevents later on
performing a reorg that replaces the tip block with an identical block
recorded with the sidechains.