Fix reorg transaction loss
There is a longstanding bug that causes the wallet to lose transactions if
they appear in a mined block, that block is reorged out, and a later reorg
reinstates the original block back in the best main chain.  This occurs due to
how relevant transactions are passed around during reorgs.  Each syncer is
responsible for providing the relevant transactions during the call to
ChainSwitch, but the syncers are not aware of the transactions that are rolled
back from previous reorgs.

Rather than passing in relevant transactions from the syncer, record these
transactions in each block node in the sidechain forest.  During a reorg, save
the removed transactions along with each block node when saving the
sidechains.