Fix sdiff used for unmined ticket pruning
Unmined tickets are removed when the required ticket price changes,
invalidating them for future blocks.  Wallet makes an attempt to remove these,
but may use the incorrect sdiff value (the sdiff from the current best block
header, not the next required block) when it is at the end of the current
ticket price window.  Calculate the correct value (which is only possible once
DCP0001 is activated), or otherwise do not prune these unmined tickets at all
using the potentially incorrect value.