ticketbuyer: Single call to PurchaseTickets.
In the case that mixing is enabled, ticketbuyer would purchase multiple
tickets by calling PurchaseTickets multiple times. This was done to
ensure that each ticket is purchased using a new cspp server connection.
This is no longer necessary since mixing moved from a client-server
model to a peer-to-peer model; a single call to PurchaseTickets is
adequate.