dcrd: Add JSON marshaling/unmarshaling tests
This tests the JSON helpers for the hashes, hashesContiguous,
headers, and transactions types.

These tests would have detected an unhandled error bug which was
recently fixed, and they will prevent a regression or similar issues
arising in the future.

The tests for each of these types are almost identical/duplicated, as a
result of the code under test being almost identical/duplicated.