[release-v2.1] Fix loop iteration broken by modernize linter
8bcc7f5a21411f939ac52744357950182b02539a rewrote the iteration over
splitPoints from a traditional for loop to range, where the i variable was
already declared, and checked following the loop.  This is a change in
semantics as the range will not set i to the length of the array, and the code
would continue with count=0 in situations where it previously would have
returned a specific error for not having any possible split denomination.

Rewrite this code in a way that the modernize linter will not break it in the
future.  In all cases where i previously would have set to the length of the
splitPoints, count is 0, so check that following the loop instead.

This fixes "message count must be positive" errors in logged by mixclient.

Backport of dc32d6e59b1ea6b0b3c6d0f74e3a8841570b1662.