Add optional badger database driver backend
This is meant for initial testing and performance evaluation, and should not
be used in production.

The addition of the badger driver requires modifications to the walletdb
interface, so this requires an eventual bump to the major module version.

The driver may be selected with the --dbdriver=badgerdb flag/config option.