Add stringer to go tools.
The repo has a dependency on the stringer executable being available on
the system path, however it is somewhat hidden because go generate is
not run as part of the regular dev/CI build.

Adding stringer to go.mod tools enables go generate to run without
having the tool installed on the path, and also enables versioning of
the tool.